+++ common/src/main/java/net/phantompig/soy/SoyConfig.java
@@ -10,16 +10,22 @@ public class SoyConfig {
 
         public static ForgeConfigSpec generateConfig() {
             ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
-            builder.comment("Enables/disables injecting entities other than the player holding the injection.");
+
+            builder.comment("Enables/disables injecting entities other than the player holding the injection");
             CAN_INJECT_OTHERS = builder.define("general.canInjectOthers", true);
+
             builder.comment("Enables/disables dispensing raw iceburst from a dispenser.");
             CAN_DISPENSE_ICEBURST = builder.define("general.canDispenseIceburst", true);
+
             builder.comment("Enables/disables titans exploding blocks upon shifting. Be warned, disabling this may lead to shifters suffocating in the ceiling immediately after shifting");
             EXPLODE_ON_SHIFT = builder.define("titan.explodeBlocksOnShift", true);
+
             builder.comment("Enables/disables titans exploding blocks upon falling from a height");
             EXPLODE_ON_FALL = builder.define("titan.explodeBlocksOnFall", true);
+
             builder.comment("Enables/disables titans exploding blocks upon attacking");
             EXPLODE_ON_ATTACK = builder.define("titan.explodeBlocksOnAttack", true);
+
             return builder.build();
         }
 
@@ -39,6 +45,29 @@ public class SoyConfig {
         }
         public static boolean shouldTitansExplodeBlocksOnAttack() {
             return EXPLODE_ON_ATTACK.get();
+        }
+    }
+
+    public static class Client {
+        public static ForgeConfigSpec.IntValue RIGHT_HOOK_KEYCODE, LEFT_HOOK_KEYCODE;
+
+        public static ForgeConfigSpec generateConfig() {
+            ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
+
+            builder.comment("The keycode for shooting the right ODM hook. If the keycode conflicts with a keybind (from the Controls menu), that keybind will be cancelled");
+            RIGHT_HOOK_KEYCODE = builder.defineInRange("client.rightHookKeycode", 69, 0, 127);
+
+            builder.comment("The keycode for shooting the left ODM hook. If the keycode conflicts with a keybind (from the Controls menu), that keybind will be cancelled");
+            LEFT_HOOK_KEYCODE = builder.defineInRange("client.leftHookKeycode", 81, 0, 127);
+
+            return builder.build();
+        }
+
+        public static int getRightHookKeycode() {
+            return RIGHT_HOOK_KEYCODE.get();
+        }
+        public static int getLeftHookKeycode() {
+            return LEFT_HOOK_KEYCODE.get();
         }
     }
 }