+++ common/src/main/java/net/phantompig/soy/client/SubjectsOfYmirClient.java
@@ -97,8 +97,6 @@ public class SubjectsOfYmirClient {
         OverlayRegistry.registerOverlay(SubjectsOfYmir.MOD_ID + "/titan_exhaustion", SoyOverlays::renderExhaustionOverlay);
         OverlayRegistry.registerOverlay(SubjectsOfYmir.MOD_ID + "/all_hardening", SoyOverlays::renderAllHardeningOverlay);
 
-        SoyKeyMappings.init();
-
         SubjectsOfYmir.LOGGER.info("Subjects of Ymir initialized on the client");
     }