+++ changelog.md
@@ -6,6 +6,7 @@ Beta
 
 - Added odm animation for falling
 - Added odm components to loot tables
+- Client config for changing odm hook keybinds
 - Fixes, mostly odm stuff
 - Fixed uniform jacket without an emblem having a blank emblem
 - New icon